Start clean-up.
